Home Assistant: Restricted Zone Notifier Python
This application [0] is designed to detect the humans present in a
predefined selected assembly line area. If the people enters
the marked assembly area, it raises the alert and sends through
mqtt. It is intended to demonstrate how to use CV to improve
assembly line safety for human operators and factory workers.

https://github.com/intel-iot-devkit/restricted-zone-notifier-python

Signed-off-by: Abraham Arce <xe1gyq@gmail.com>